+++ source/renderer/app/components/wallet/transactions/Transaction.js
@@ -9,7 +9,7 @@ import { Link } from 'react-polymorph/lib/components/Link';
 import { LinkSkin } from 'react-polymorph/lib/skins/simple/LinkSkin';
 import CancelTransactionButton from './CancelTransactionButton';
 import styles from './Transaction.scss';
-import TransactionTypeIcon from './TransactionTypeIcon.js';
+import TransactionTypeIcon from './TransactionTypeIcon';
 import adaSymbol from '../../../assets/images/ada-symbol.inline.svg';
 import arrow from '../../../assets/images/collapse-arrow.inline.svg';
 import {