+++ source/renderer/app/config/stakingConfig.js
@@ -1,15 +1,15 @@
 // @flow
 import type { RedeemItnRewardsStep } from '../types/stakingTypes';
 
-export const RANKING_SLIDER_RATIO = 60;
-export const MIN_DELEGATION_FUNDS_LOG = Math.log(10);
-export const MIN_DELEGATION_FUNDS = 10;
 export const MIN_REWARDS_REDEMPTION_RECEIVER_BALANCE = 1; // 1 ADA | unit: ADA
-export const INITIAL_DELEGATION_FUNDS_LOG = Math.log(1000);
+
+export const RANKING_SLIDER_RATIO = 60;
+export const MIN_DELEGATION_FUNDS = 10;
+export const MIN_DELEGATION_FUNDS_LOG = Math.log(MIN_DELEGATION_FUNDS);
 export const INITIAL_DELEGATION_FUNDS = 1000;
-export const MAX_DELEGATION_FUNDS_LOG = Math.log(220000000);
-export const MAX_DELEGATION_FUNDS = 220000000;
-export const OUT_OF_RANGE_MAX_DELEGATION_FUNDS = 33000000000;
+export const INITIAL_DELEGATION_FUNDS_LOG = Math.log(INITIAL_DELEGATION_FUNDS);
+export const INITIAL_DESIRED_POOLS_NUMBER = 150;
+export const CIRCULATING_SUPPLY = 31736470500;
 export const ALL_WALLETS_SELECTION_ID = '0';
 
 // Dimensions