+++ source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsConfirmationDialog.js
@@ -211,8 +211,7 @@ export default class DelegationStepsConfirmationDialog extends Component<Props> 
             hwDeviceStatus !==
               HwDeviceStatuses.VERIFYING_TRANSACTION_SUCCEEDED) ||
           isSubmitting ||
-          !transactionFee ||
-          !!error,
+          !transactionFee,
       },
     ];