+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 ## vNext
 
+### Features
+
+- Implemented user configuration for the SMASH server ([PR 2259](https://github.com/input-output-hk/daedalus/pull/2259))
+
 ### Fixes
 
 - Fixed the overlap of the "X" button in the stake pools search box ([PR 2251](https://github.com/input-output-hk/daedalus/pull/2251))