+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Changelog
 
 ### Fixes
 
-- Fixed visual glitch on the transaction list switching between filters ([PR 2249](https://github.com/input-output-hk/daedalus/pull/2249))
+- Fixed visual glitch on the transaction list switching between filters ([PR 2249](https://github.com/input-output-hk/daedalus/pull/2249), [PR 2261](https://github.com/input-output-hk/daedalus/pull/2261))
 - Fixed Newsfeed drop shadow when there is an update item ([PR 2242](https://github.com/input-output-hk/daedalus/pull/2242))
 
 ### Chores