+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 ## vNext
 
+### Features
+
+- Close Newsfeed when clicking anywhere else in the UI ([PR 2250](https://github.com/input-output-hk/daedalus/pull/2250))
+
 ### Fixes
 
 - Fixed empty disk space calculation for some Linux distributions ([PR 2258](https://github.com/input-output-hk/daedalus/pull/2258))